Moved check_version from `pipeline/tools/check_version.py` to `pipeline/core/check_version.py` Moved test for check_version to appropriate location.
