+++ pipeline/tools/__init__.py
@@ -2,4 +2,3 @@ from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
 from . import reference_lamp_factory
-from . import check_version
